Rewrite TermValidatorFactoryTest and add missing test cases
I'm introducing a new test here that passes the same strings to all three
label/description/alias validators and expects them all to do the same for
the provided strings.

Then I realized that the existing test functions are all identical, and do
not add value to the new test I just wrote. So I removed all redundant
tests.

The new test covers much more than the old, e.g. the fact that spaces are
allowed, but tabs are not.

Change-Id: Ie26e50f5a065e9488cabe8f844edf8d12b84026a