Update patch set 2
Patch Set 2: Code-Review+1

(1 comment)

The code in this patch looks fine for me.

This really should not be an option. These options are meant to be exposed to API users. In some cases we are misusing these options to pass in arrays or objects – something the user can not do. But the option in question here really should be a separate constructor parameter.

Do not forget to add the new setting to docs/options.wiki!

Patch-set: 2
Label: Code-Review=+1