Update patch set 2
Patch Set 2:

I said this before: From a users perspective, the things in the Data: namespace *are* files. They just happen to be in an other namespace, and stored in an other table.

However, in code, comments, backend documentation and so on we should make sure to *not* call it "files" to avoid confusion with what MediaWiki core calls "files".

Patch-set: 2