Update patch set 42
Patch Set 42: Code-Review+1

(7 comments)

Full review done. Phew, this is huge. But a very lovely diff. I can literally see how much work you put into this. This is awesome! Thank you so much for doing this!

I wish this patch could be split into multiple smaller ones. It's always hard to review patches with +/-500 lines (even if this removes 100, which is great to see). But it probably can't be split because all these are subclasses of SpecialModifyEntity.

It also scares reviewers to see a patch with more than 40 patch sets, even if I can understand why that happened. Again, splitting it into smaller patches (even if you can only split of a few lines into separate patches) would help a lot.

As all my comments this is not meant as an edit request, just a remark. I will merge this as it is. All I'm asking for is a bit of input from you, as well as Aleksey.

I tested this by actually trying out all changed forms, one browser tab with this version, the other with the current version on https://wikidata.beta.wmflabs.org.

One detail makes me wonder: there is sooo much whitespace between the input fields, it almost appears like they do not belong together. This is especially bad in Special:SetLabelDescriptionAliases. Is this just how OOUI is designed?

Patch-set: 42