Update patch set 1
Patch Set 1:

> […] should be, probably, merged with existing EntityChangeTest

That exactly is the plan. I just wanted to have a sequence of small patches instead of a big mess nobody wants to review. I will continue working on this, but I will not invest more time on this before these baseline patches are merged.

> ChangeRowTest now tests EntityChange but not the abstract class […]

Yes and no. Both the tests and the tested code are still the same. The only difference is that the test does not *know* any more where the tested code is. As before, this is nothing but an intermediate step. I want both the abstract class and it's tests to die.

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0