Refactor foreign EntityId parsing in LinkBeginHookHandler
Main reason to touch this code was the fact that two methods implemented
the string "Special:EntityPage/".

I also removed the "lookup as a local ID using the $this->entityIdLookup
service" from the method, and moved this out. What's left is a method that
exclusively deals with foreign IDs.

This should not change any behavior, this is pure refactoring.

Bug: T157004
Change-Id: I26893ee1c0bc0cc26b5c38db5c56ec271e059c56