Update patch set 1
Patch Set 1: Code-Review+1

So there would be a single instance of this class for each settings that should be exposed in mw.config?
Makes sense. I only wonder whether proper English wouldn't be "SettingValueProvider", especially if this is supposed to provide a value of a single setting.

Patch-set: 1
Label: Code-Review=+1