Update patch set 9
Patch Set 9:

Sorry for getting back to this after a week. Merged patches are not that visible on my gerrit board.

 > That's why I did not suggested something like "Foreign" or "RemoteRepositories", but "Federation". This word does not mean it's about foreign repositories only.

Well: "federation" of repositories means that there is more of them than one. It does not mean indeed they have to be a local repo plus some other ("foreign") ones. It could, for needs of theoretical discussion, be also many foreign repos not including the "local" one. Why not.
But this all does not quite match with the purpose of this class. How I think of it, it is supposed to be a general object providing repository settings (be it local or foreign repo, doesn't matter) to all other Wikibase objects that need them. In other words, the idea is to use this in places that used to directly access setting X or Y from SettingsArray object.
It is true that services we've introduced for "federation" made this need more evident (as we now cannot hard code the setting for database name, for example), and I have biased the focus in this direction with part of the commit message, but I believe this class should not be seen as the federation-only class.

 > Or the other way around: Without federation this definitions thingy is not needed, not even for the own local repository.

See above. And this "thingy" is not necessarily needed for federation either. It is expected to make plenty of things easier and less fragile. But we could deal without it possibly, as the "federation" related code used to do so far.

Patch-set: 9