If search term is a entity ID add known repository prefixes
If entity ID is given, prefixes of repositories configured
as providing entities of the given type are prepended to
the given search term, and a resulting entity ID is search
in respective repositories.

In particular, it is possible to search for entity ID
provided by the foreign repository without providing
the repository prefix in the search term.

If multiple repositories provided entity of the given type,
and accordingly prefixed matches were found in several
repositories, all relevant matches will be returned.
As of January 2017 single entity type is only provided by
single repository but the code in EntitySearchHelper
would be able to handle the multi-repo case if this
was changed.

Bug: T158169
Change-Id: I2405dcdc2cedaf17b1df363bf7ea2918a338b909