Trivialize InterWikiLinkWikitextFormatterTest
Main change in here is the test for spaces and single quotes. This was
not fully tested before. Spaces should not be encoded in the label. Quotes
should be encoded in the label, but must use an other encoding.

This change also changes many self:: calls to phpunit assert methods to
call $this-> instead.

Bug: T57549
Change-Id: I94309aaa3408b78dc625eba7677c361bf93120e2