Update patch set 10
Patch Set 10:

I ended doing much more in my follow up, also touching files that are not part of this patch. I may do some changes here, but want to do this only if one of the issues actually blocks this patch from being merged. Do you think this is the case? Then either set a -1 or +2, please.

Patch-set: 10