Update patch set 2
Patch Set 2:

Thanks!

Having another patch that uses this would be good, so we can try this out manually, and see how it integrates with existing code.

For reviewing individual bits of code, small patches are good. For experimenting with new approaches, it's often necessary to have a chain of patches that actually shows how the new stuff is going to be used.

Patch-set: 2
Label: Verified=0