Remove EntityChange::__toString
I feel silly because Aude uploaded the exact same patch months ago, and
I rejected it.

Reasons for this removal: It's a pretty complex function, but not used.
It binds against a bad implementation detail (the internal fields array)
I want to get rid of. Simply removing this makes following patches much
easier.

Change-Id: I64d49ea94681d25531465b52aecf21306ac7e2ac