Update patch set 11
Patch Set 11: Code-Review+1

This looks good to me. We|ll find out if it actually works when we hook it in ;)

No +2 yet because I wonder... why inject the validators? Wouldn't static composition be fine in this case? Why factor them out at all? Usually, the deserializer would validate on the fly. Why split that?

Patch-set: 11
Label: Code-Review=+1