Update patch set 2
Patch Set 2: Code-Review+2

Manually tested. Works. The new design is really lovely.

One detail I found is that the original edit form does not have a pipe between the save and the cancel button. I believe we can remove this pipe here. But it was there before and not the intention of this patch to touch this detail.

Patch-set: 2
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2