Update patch set 12
Patch Set 12: Code-Review+1

(2 comments)

Please have a look at what I did in PS12. I removed the check from ValueViewBuilder. I'm aware this means the code is not very clean, but I believe we can easily live with this "acceptable debt". Note that no other constructor parameter is checked. I find this acceptable because there is a single constructor call in production code, and the parameter is guaranteed to be set in this constructor. We don't make the code so really better by checking all these parameters twice.

Patch-set: 12
Label: Code-Review=+1