Update patch set 2
Abandoned

Using the composer only makes sense if either the entity type is variable, or the repo prefix is (or if there is a chance we will make one of it variable in the near future). Neither is the case here. We know the repo prefix is empty. We know the type is "item". Nothing wrong with using the static constructor then. The "wb_items_per_site" table that is used here (via SiteLinkTable) does not support repo prefixes, and as far as I know it does not need to.

Patch-set: 2
Status: abandoned