Update patch set 2
Patch Set 2:

I still like this patch ("still" because it's almost a month old now). What I love is that this entirely removes two global variables.

I was wondering about the two includes, but I believe this does not mean there is any new dependency. These includes happened before, just in an other file, but still in the same components.

I need to play around with this patch in my dev system before I can merge it.

This also seems to conflict with If559e86. We should make sure the two patches are not making the code a mess when both are merged. Daniel, can you have a look at this here?

Patch-set: 2
Label: Verified=0