Do not use SerializerFactory::newEntitySerializer in tests
It's not wrong to use this method, but it's not necesarry, and the name
of the method is confusing. This is not a generic entity serializer, but
a serializer for items and properties only.

This is very similar to using ItemIdParser in tests when a test does not
need to know about other entity types.

Bug: T160426
Change-Id: I02571a90b1e1d9418a137dc5af5644e0a396aac6