Update patch set 4
Patch Set 4:

> is not really fixing a bug linked in the commit message […] Bug is already fixed

This is a direct follow up for what was done in T160426. I think it's very useful to link it to the same ticket, but I do not think it would be good to reopen it. This is just usual refactoring.

Patch-set: 4
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0