Update patch set 1
Patch Set 1:

You're right about 2. It should be rephrased so it doesn't refer to classes that do not really exist but explains what it is meant here. But this is something not to be changed here IMO.

Iaf79a0012e691445751644a194a67b8ae14fba1e introduces PropertyRdfBuilder so what you expect in point 1 would be somewhat starting to get fulfilled soon.

Patch-set: 1