Update patch set 2
Patch Set 2:

Is there a way to trigger an actual wrong behavior with the old code? I tried but can barely think of a way. I believe the only way to make this not do what it is supposed to do is to add HTML to one of the two messages. Is this correct?

But then again, when is this code path triggered?

Patch-set: 2