Introduce providers for 4 more data types in ValidatorBuildersTest
All this patch does is moving existing test cases around, but does not
change any detail.

I'm introducing 4 more dedicated test case providers for 4 data types.
The idea is to make the giant provider that was there before much smaller
and easier to maintain. (And probably turn it into an "mismatching data
type" provider if only cases with mismatching data types are left.)

Note that I'm turning the last element in the arrays into array keys.
This is not possible with the giant array because there are a lot of
duplicate messages in there.

Change-Id: Ieef40557a77e2b8dabf19578c7e66f89c4252027