Update patch set 4
Patch Set 4: Code-Review-1

(7 comments)

I like the idea of having this class. I think we should continue working on this patch and get it into a mergeable state.

But I'm afraid this does not solve the issue we want to solve. Not even remotely. All this patch does is moving some code around, from places where Repo and Client intersect to a new place where Repo and Client intersect. This is a good move, because it makes these places a bit more visible. But I believe we need to do something very different to actually untangle the cross-dependency we discussed.

I strongly suggest to move the new file to the lib/config/ directory, and not to a namespace in the includes/ directory. It's not really a class and does not belong there.

Patch-set: 4
Label: Code-Review=-1