Update patch set 4
Patch Set 4:

> I strongly suggest to move the new file to the lib/config/ directory, and not to a namespace in the includes/ directory. It's not really a class and does not belong there.

I vote against moving it to `lib/config`. It is class (static one, but class) and will be autoloaded. Otherwise we will have to introduce yet another autoload rule to load this single thing.
Also it does not match other things in `config` directories at all, which are config examples.

Patch-set: 4