Update patch set 4
Patch Set 4:

(5 comments)

But I'm afraid this does not solve the issue we want to solve. Not even remotely.

> No - it removed the dependency that client used to have on repo. It does not remove the opposite dependency. It turns the cross-dependency (very bad) into an unwanted optional dependency (still bad). 

> All this patch does is moving some code around, from places where Repo and Client intersect to a new place where Repo and Client intersect. 

That's what refactoring is - moving code around. With this move, Client no longer has a dependency on Repo. I believe that is an improvement.

> I strongly suggest to move the new file to the lib/config/ directory, and not to a namespace in the includes/ directory. It's not really a class and does not belong there.

I agree that it would be nice to split bootstrap code from application logic. And I agree that this code should not live in lib. I do not think it should be in config, though, for the reasons stated by Aleksey.

Patch-set: 4