Avoid database access in ItemHandlerTest data provider
ItemContent constructor access the database when a redirect title
is passed. (to check if the redirect has same content model
and the title doesn't already have content model set on the object)

We can work around this by setting the content model on the
title object.

It would be nicer if we could altogether avoid database access
in the constructor (maybe by removing the check and checking
elsewhere when creating redirects?) but not sure and don't want
to do that in this patch.

Change-Id: I16c3252892595ef79a9f865c850e02066093d053