Update patch set 5
Patch Set 5:

As said I believe this patch is an improvement and should go in.

> With this move, Client no longer has a dependency on Repo.

This statement keeps confusing me. All this patch does is moving the dependency to a class in the libs folder. Client still uses this class, so client still depends on repo. This patch does not change this.

> It turns the cross-dependency (very bad) into an unwanted optional dependency (still bad).

This might provide an alternative view on what I wrote above, but I'm afraid I don't get it.

Patch-set: 5