Update patch set 11
Patch Set 11:

For the reviewer's benefit, please leave the namespaces unchanged in this patch, and change them in another patch. That makes it easier to see what is going on.

Similarly, it helps to split any changes to factory logic from the patch that moves the files.

Patch-set: 11
Reviewer: Daniel Kinzler <128@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0