Update patch set 2
Patch Set 2:

Re custom entity types: right now, we have one special case: Items have sitelinks, and when sitelinks change, we have link/unlinnk events on the  respective client. The EntityChange has to contain the relevant data (e.g. the site diff, or a compact variant of it). Otoh, we don't want all changes to contain the full diff (we had that, it was causing performance issues). Do we need at least some knoweledge about what "extra" information is needed for which entity type, and we need a place to put that information.

Or we say that sitelinks are the only extra we will ever support.
I currently do not think we need anything extra for Lexemes and MediaInfo. But maybe I'm wrong - and it would be sad if we had to completely refactor this again just to make room for some extra bit of info we need to loop through change propagation.

Patch-set: 2