Update patch set 11
Patch Set 11:

> For the reviewer's benefit, please leave the namespaces unchanged
 > in this patch, and change them in another patch. That makes it
 > easier to see what is going on.
 > 
 > Similarly, it helps to split any changes to factory logic from the
 > patch that moves the files.

This patch is not supposed to be merged. I did it to discover the tree of classes that should be extracted to get understanding what is there and what for. Also to understand what parts of the code depend on those.
Changing namespaces helped me a lot to achieve this.

I uploaded this patch for everyone to see what this tree is.

Patch-set: 11
Reviewer: Aleksey Bekh-Ivanov (WMDE) <4089@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0