Update patch set 5
Patch Set 5: Code-Review-1

(5 comments)

There are some small problems. Should be easy to fix.

I hacked it up locally, and the test passes for the SQL based locking. But it also passes with NullLockManager - which means the test doesn't actually work... Have you been able to make this fail when not using any actual lock? If we can't do that, we can't really test locks.

Patch-set: 5
Label: Code-Review=-1