Remove incompatible floating point numbers from all RDF tests
The issue is that 12.345 is a different thing on 32 and 64 bit PHP as
well as HHVM. The hash calculation of GlobeCoordinateValue uses these
floating point numbers and creates different hashes on the different
systems, even if the input on all is 12.345.

This patch does not fix this issue. The issue needs to be fixed by
changing the serialization (or at least the hash calculation) of
GlobeCoordinateValue. This patch makes tests that should not care about
this issue not fail on the different systems, because ints behave the
same.

Bug: T161231
Change-Id: I39185f3db3b2cc343edcd384c900686cd46db19a