Update patch set 4
Patch Set 4:

I'm not sure if this is the right place to discuss, and also do not understand what you mean. Are your comments meant as edit requests? What should I do then?

https://github.com/wmde/WikibaseDataModelServices/pull/157 in fact "encapsulates" the code that uses the clear method.

But this is actually unrelated to what this patch does. It removes the only (!) place in our entire code base that fails hard with an exception if an entity type does not implement the holder interface. As already said this is an actual blocker right now.

Patch-set: 4
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0