Update patch set 2
Patch Set 2:

As I said multiple times in If129df5 and in the IRC chat we just had I would not add any other file than TermValidatorFactory.php. It doesn't hurt much to do this, but is not needed to fix the problem in production. I had the impression it's good practice to keep backports as small as possible. I will leave this to the people who have merge rights here to decide.

Patch-set: 2