Update patch set 3
Patch Set 3: Code-Review+1

Aude, I totally trust you that these tests are actually hitting the database. But since you set a -1, I wonder what your plan is? I also wonder if the Database tag is really needed if a test does not write anything to the database. E.g. it appears the template module reads messages, others probably read the sites table, but do not write to the database.

In short: We should have a plan to make more tests independent from the database. It would be cool if you could explain the most easy way to find good candidates, and how to run them in a way I can easily see if the database is hit or not.

Patch-set: 3