Always escape user input in UserInputException
I played around with different possible ways to fix this. I started
adding wfEscapeWikiText to all calls of the UserInputException as well
as MessageException constructors. Then I realized that *all*
UserInputException have the exact same problem. I decided to change the
interface of the class to expect unescaped user input (this fits very well
to the name of the class), and do the escaping in the constructor.

The bug describes an attack vector that allows injecting arbitrary
Wikitext. It does not allow injecting arbitrary HTML or JavaScript.

This patch is intentionally minified, so we can easily backport it.

Bug: T161418
Change-Id: If6e79745ec79ba3574694601a0a75fd93f0246a8