Update patch set 2
Patch Set 2: Code-Review-1

It looks like the two shades of blue are flipped in the 2 snaktype as well as 2 rank selector images. Before it was dark and had a brighter hover effect. Now it's bright and does have a darker hover effect. I believe this was a mistake. It means the default icon is much brighter, which I think is not a good idea.

Patch-set: 2
Label: Code-Review=-1