Consolidate RDF test data.
This patch removes redundancy from the RDF test data by splitting the test files
into re-usable parts.

The intent is to reduce the effort needed to adjust test data files when extending
or changing the RDF binding.

REVIEW NOTE: Note that no application logic changed, and the unit tests only changed
in how they load the test data. Since the actual output has not changed and the tests
still pass, a line-by-line review of the test data should not be needed.

Change-Id: I340c9bfa96160606dc828e69984e82a00d8da9e3