[WIP]Consider foreign entities in RDF mapping
Namespaces used in clauses related to foreign entities now
contain the repository name (prefix) of the repository
the entity belongs to.

Repository prefix is also dropped from the "local name" of
the foreign entity.

In order to correctly build both concept and description
URIs, base of the description URI also became a part
of the repository definition.

TODOs:
 - how to set default canonical data URI base? New setting?
 - split changes to RepositoryDefinitions out
 - fix CS issues

Bug: T161592
Change-Id: I6acb00b23425bdce4dd03227415e593595b3c1cc