[WIP]Consider foreign entities in RDF mapping
NOTE: this is half-done patch, demonstrating the possible
approach but lacking many details being implemented yet.

Namespaces used in clauses related to foreign entities now
contain the repository name (prefix) of the repository
the entity belongs to.

Repository prefix is also dropped from the "local name" of
the foreign entity.

In order to correctly build both concept and description
URIs, base of the description URI also bacame a part
of the repository definition.

TODOs:
 - have rdf data uri base configurable by RepoDefinitions
 - split changes to RepositoryDefinitions out
 - missing unit test for new method of RdfVocabulary
 - any more RdfBuilder test cases needed?
   - property predicates
   - prop novalue
   - more?
 - failing RdfBuilderTest::testAddEntityForeignEntity
 - more?

Bug: T161592
Change-Id: I6acb00b23425bdce4dd03227415e593595b3c1cc