Update patch set 1
Patch Set 1: Code-Review+2

I'm merging this to allow us to experiment. But I suggest to not create a subclass of this. This base class is so trivial, we do not win much by subclassing it. Instead create a new class with exactly the code we need.

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2