Update patch set 1
Patch Set 1: Code-Review+2

Code looks good so far. I think this is a good move, but I have one concern: This introduces a method that does have the same name and overall signature in different traits, but different additional guarantees. One requires getEntity() to return a LabelProvider, the other a StatementListProvider, and so on. Is this good practice? It looks like this might become confusing very fast.

A very closely related problem is that there is no way to express an "and" in a PHPDoc tag. The current "StatementListProvider|EntityDocument" type hints are a hacky workaround. I'm totally fine with this workaround, I just wanted to point this problem out.

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2