Update patch set 3
Patch Set 3:

(2 comments)

Can the unit test for this API module now be improved? We can now mock all services, so we no longer need to mess with global state and databases...

Doesn't have to happen in this patch, but i'd want a follow up or at least a ticket for that if it doesn't happen here.

Patch-set: 3
Label: Verified=0