Update patch set 8
Patch Set 8:

(1 comment)

The issue is that getEntityRdfBuilderFactory() caches the EntityRdfBuilders. If they have been created during an earlier unit test, they will be using a different SiteLookup. 

The solution is to use a fresh instance of WikibaseRepo in each test run. This only works if no code used in the test calls WikibaseRepo::getDefaultInstance.

I'll amend the patch.

Patch-set: 8
Label: Verified=0