Update patch set 1
Patch Set 1:

(1 comment)

Since both the read and the write code have to work with the old and the new schema, we can't really rely on the read code to test the write code. You'll have to look at the database directly to check if the code behaves as expected. I think this is fine in a unit test that tests a class that talks directly to the database. Afterall, the effect on the database are part of the classes contract.

Patch-set: 1