Update patch set 2
Patch Set 2:

My current thinking is that we should declare the base data URIs for each repo, and perhaps RdfVocabulary should even know about them, even though we don't currently need them. The rationale is that the RdfVocabulary should provide all information needed to represent data from or in any of the known repos, and should not know or care about which info about which repo is used when or where.

I suggest to open a ticket, so we have a to discuss this, where the discussion can later be found.

Patch-set: 2
Reviewer: Daniel Kinzler <128@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0