Update patch set 4
Patch Set 4: Code-Review-2

I think we need to investigate and discuss this a bit more before doing such a significant change to one of our absolute basic building blocks. Setting a temporary -2 for the moment. Not meant to block this patch, the contrary: to discuss, understand and agree.

Patch-set: 4
Label: Code-Review=-2