Update patch set 6
Patch Set 6: Code-Review+1

I talked with Aleksey and understand the issue much better now. This is indeed fixing a bug. In PS6 I fixed a remaining inconsistency: the code mixed prop() and attributes.

I also tried to implement this natively without the additional dependency, but decided it's worth to have it.

Patch-set: 6
Label: Code-Review=+1