Update patch set 4
Patch Set 4:

Yeah, I just chatted to Daniel and there will be different flag, or maybe even no flag, so nothing to do here.

Final thing that came to my mind: maybe we want to have some TermSqlIndex unit test with a flag on? Test's going to be a bit silly but it would be nice to have it, if this is not much of hassle, so the patch is kind of "complete".

Patch-set: 4
Label: Verified=0