Support EntityIds in diff serialization in EntityChange.
This adds support for serializing diffs that contain EntityIds to EntityChange.
So far, EntityChange has support for serializing Statement objects in diffs,
but not stand-alone EntityIds, as are needed to support Lexemes.

NOTE: this touches on a very ugly bit of code, and adds more uglyness.
This should be improved by introducing DiffSerializer / DiffDeserializer
classes, and pulling serialization logic ot of EntityChange, into the
ChangeFactory.

Change-Id: I4430dc7ade00b75d91a08be29f60e54d20aff7e2
Needed-by: If46efb61a0abacc7632b125643eb749946143785