Update patch set 3
Patch Set 3:

@Ladsgroup: maybe you could have a look at If686b79e70197b70c4a73545664eb7f0f0dd7a6c where I got somehow tests to pass.

What I did is overriding the content handler for "wikibase-item" content model for the purpose of the test, so it uses the "custom" item handler.
I would have expected this to be enough. But when I just moved the new test to the separate method, so that modifying the global does not mess with other tests, the test for $cases['claims'] (which should be completely unrelated to testFoo) started to fail as wb-identifiers was 1 instead of 0. This is a bit odd to me. It looks like this test also used the "custom" item handler. setMwGlobals should be making the custom value of the global only be available in the particular test calling this method, and restore the "default" value after the test, so I don't quite get what happens there. Maybe Daniel has an idea?

Patch-set: 3