Update patch set 1
Patch Set 1:

Thanks for merging this. As mentioned in the commit message this is an actual requirement for possibly removing isValid. But I believe the change here is still useful, independent from said discussion.

Serializing an entity usually should not fail, but there is no guarantee. The fact that EntityContent::isValid returns true is no guarantee the serialization succeeds. That's when this try-catch becomes relevant.

I believe we can not reproduce this edge-case in production currently, but this might change with every new entity type and every new serializer. That's why I think this try-catch should be there.

Patch-set: 1