Update patch set 3
Patch Set 3:

One solution is to move getEntityPageProperties to EntityHandler. It can just take the EntityContent as a parameter.

In general, Content objects do way too much, they often need services, which they can only get from global state, or have injected from the ContentHandler. This is my fault, I got this wrong when I defined the Content interface five years ago. 

The recommendation is to push as much logic as possible into ContentHandler. 

That refactoring would however go far beyond the scope of this patch.
As a quick fix, perhaps the getContentHandler() method can be mocked on the object under test. That's a bit hack, but could work.

Patch-set: 3
Label: Verified=0