Update patch set 2
Patch Set 2: Code-Review-1

(3 comments)

Important question: How bad is it if we change this serialization format later? This makes a big difference for reviewing this. If we can easily change it later, I'm even willing to +2 this. If it's not, I would like vote with a temporary -1 to be sure my suggestion (see inline comments) is seen and considered.

Also note that nothing of this is necessary when we design If46efb6 in a way that it does *not* use EntityId objects, but sticks to ID strings instead (which is the status quo already).

Patch-set: 2
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1