Consider languages being it's own parent in LanguageFallbackChainFactory
As far as I can tell this change is not that relevant in production because
we are currently exclusively using the …::FALLBACK_ALL mode in production.
But I feel this makes the code more clean and behave more like you would
expect: if you don't ask for a self-reference, you should not get one, even
if it is in core's variants list for some reason.

Bug: T156280
Change-Id: I1d9691c784032669979f8109c9a5f65cbf4122c9