Update patch set 2
Patch Set 2: Code-Review-1

I really think we should stick to the name that is already there and not invent a new one. This is confusing. "Base URL" could mean a lot of things. Note that we want, in theory, two URLs for geo-shapes: one links to the MediaWiki page, the other links directly to the JSON file without any MediaWiki chrome. The later does not exist at the moment.

The problem is: "base URL" sounds like the later.

The word "frontend" does make it more obvious it links to the "file description page" of the geo-shape.

Patch-set: 2
Label: Code-Review=-1