Update patch set 3
Patch Set 3: Code-Review+1

(1 comment)

I reviewed all changes individually, and I believe they are all fine. Thanks for working on this. I believe we can +2 this as it is, but I'm not doing this right away to give my colleagues some time to become aware of this.

As far as I understand not all rules are enabled at the moment. Is this correct?

Patch-set: 3
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1