Update patch set 5
Patch Set 5:

(4 comments)

Thanks Daniel for the hint on ContentHandler instance caching. I missed that.

The patch looks good now IMO. Few minor suggestions in ItemHandlerTest (names) but I'd be OK with merging this after the indentation in the test class is fixed.

Indeed, it is a bit hacky now, so further refactoring should happen but I agree this should not be done in this patch so it does not grow to some enormous size.

Patch-set: 5
Label: Verified=0