Update patch set 3
Patch Set 3: Code-Review-1

(6 comments)

Full review done. No big issues, just a few tiny details that should be very easy to fix.

Note that this will enable the new type right away. We might want to add it to the "disabled" setting for a while. Can you create a configuration patch for this?

Patch-set: 3
Label: Code-Review=-1