Transform user ID from repo to client in change prop
On the repo side, save the user's central user ID (if they are part of
a central login system) into the change.

On the client side, transform this back to the user ID for the client
wiki.  If they have not logged into the client wiki, it is not part of
a central system, etc., it will be transformed to 0.

Plus minor typo fix in ChangeHandler.

Bug: T51315
Change-Id: Ie7b9c482cf6a0dd7215b34841efd86fb51be651a