Update patch set 1
Patch Set 1: Code-Review+1

This drops support for PHPCS 0.4.0 and 0.5.0 which has been there for a while. But I don't think this is a problem. That said, I am letting others to stop me from merging this if there are an negative implications of this change I am not aware of. Will merge if no one objects in a bit.

Patch-set: 1
Reviewer: WMDE-leszek <2626@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1