[WIP] Remove meaningless ->expects( $this->any() ) from all tests
WIP because I would like to do the same for all tests, but start with a
trivial directory to see if this is feasible.

We might disagree on this because this change makes PHPStorm complain.
It does not understand why this method exists on this interface (actually
I don't understand this myself).

Change-Id: Ib15f28c5e758297e26ed7957ba9191468c7fc05c