Update patch set 5
Patch Set 5: Code-Review+1

This sounds like the exact same issue Aleksey run into when he tried to split test-only code off into a "hamcrest/hamcrest-php" library. The only solution we found was to add it to the "require-dev" section in core. I believe we should do the same here.

Patch-set: 5