Refactor and fix HTML injection in SpecialAvailableBadges
This is a direct follow up for the code introduced in I8cd2912.
Major elements of this refactoring are:
* Fix the HTML injection when outputting unescaped $description.
* Use a message instead of a static " - " string (with the short dash,
which is a not so nice replacement character). I decided to use
"comma-separator" because this is the same as on
Special:ItemDisambiguation and other special pages.
* Concatenate a $html string instead of calling addHtml() multiple times.
This is not done for performance reasons but to make the code easier to
read.

Bug: T114473
Change-Id: Idb148e09808bba45be3b894ef361b1d9ed755a4e