Update patch set 1
Patch Set 1: Code-Review-1

I'm afraid this misses all context. Therefor the -1.

The Phabricator ticket does not explain why the setting needs to be renamed.

To what does "can be anything" refer to? Sure, thats why it's a setting, isn't it? You can set the icon to anything. This does not mean we have to rename the setting.

Finally, is this used in any production or test system? We need to update this along with this rename, and probably have both the old and the new setting in parallel for a while. This is not mentioned in the commit message or ticket.

Patch-set: 1