Update patch set 1
Patch Set 1: Code-Review+1

Code is fine, but I'm missing a lot of context. I don't find a comment talking about the number 9 in the ticket. Why not 10? Why not 100? How does it help to run a bunch of smaller queries in a loop? This appears counter-intuitive to me. I believe it would help a lot to mention the index that is used for this query, and why it performs better with such small chunks.

Patch-set: 1
Label: Code-Review=+1