Update patch set 2
Patch Set 2: Code-Review+1

This patch appears to be incomplete, because the icon is not used anywhere. We could at least mention the fact that there is a default icon in the default config and config documentation.

Patch-set: 2
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1