Update patch set 2
Patch Set 2:

Also note that the icons for Echo needed to be 30 x 30 pixels the last time I checked. The uploaded one has a very odd size.

Also note that we should aim for SVGs instead of PNGs whenever we can.

I believe all this can be fixed later, when there is code that allows us to test this.

Patch-set: 2