Update patch set 1
Patch Set 1: Code-Review+1

Direct editing like https://www.wikidata.org/wiki/Q999999999?action=edit is not possible with and without this hook. But this was not the reason this code was created. It exists to remove the edit link that appears when using Special:Search on an entity namespace, see T58032. By removing this hook this bug will come back.

If we can't find a better solution for now we should still merge this patch. T58032 is not critical, just an ugly UX issue. But I believe it's quite critical to have a test in place that makes sure we are not reintroducing the "can't restore entities" issue we are trying to solve by removing this.

Patch-set: 1