Update patch set 3
Patch Set 3:

All DropdownWidget examples on https://doc.wikimedia.org/oojs-ui/master/demos/ have a "Select one" placeholder that does *not* appear as an element of the drop-down list. How is this possible? Why can't we use this? I had the impression a core idea of this OO framework was to close the gap between PHP and JavaScript. But such significant differences mean these are two pretty much independent libraries, which is a shame.

Can we at least put the placeholder text in brackets (so it says "(pick a data type)" with the first character lower case), or make it gray?

Patch-set: 3